[PATCH] doc: Commit performs rollback of aborted transactions
The Commit command handles an aborted transaction in the same
manner as the Rollback command.  This needs to be documented.
Add it to both the reference page as well mentioning the behavior
in the official material for introducting transactions - the tutorial.

Note the our behavior is contrary to the SQL Standard as well.

In passing, make the description of the Commit reference page
self-contained by mentioning the 'and chain' behavior.