doc: Add an overview of NULL treatment in PostgreSQL
The handling of NULL within PostgreSQL is broad, varied, and
unintuitive.  This new section provides a place for the reader
to go and skim over the various behaviors and stick them in
the back of their mind.

Some existing documentation is tweaked to be more in line with
the overall flow of the material presented in this overview.
The individual features do still constitute an authoritative
location for describing the behaviors - but now have a more
convenient place to put examples.